Crate yanix

yanix stands for Yet Another Nix crate, and, well, it is simply a yet another crate in the spirit of the nix crate. As such, this crate is inspired by the original nix crate, however, it takes a different approach, using lower-level interfaces with less abstraction, so that it fits better with its main use case which is our WASI implementation, wasi-common.

Modules

clock

dir

fadvise

fcntl

file

filetime

This module consists of helper types and functions for dealing with setting the file times (mainly in path_filestat_set_times syscall for now).

poll

socket