Module yanix::filetime

This module consists of helper types and functions for dealing with setting the file times (mainly in path_filestat_set_times syscall for now).

The vast majority of the code contained within and in platform-specific implementations (super::linux::filetime and super::bsd::filetime) is based on the filetime crate. Kudos @alexcrichton!

Enums

FileTime
A wrapper enum around filetime::FileTime struct, but unlike the original, this type allows the possibility of specifying FileTime::Now as a valid enumeration which, in turn, if utimensat is available on the host, will use a special const setting UTIME_NOW.

Functions

utimensat
Wrapper for utimensat syscall, however, with an added twist such that utimensat symbol is firstly resolved (i.e., we check whether it exists on the host), and only used if that is the case. Otherwise, the syscall resorts to a less accurate utimesat emulated syscall. The original implementation can be found here: filetime::unix::linux::set_times